+++ server/utils/isValidDateOrder.mjs
@@ -1,0 +1,5 @@
+const isValidDateOrder = (startDate, endDate) => {
+  return Date.parse(startDate) <= Date.parse(endDate);
+};
+
+export default isValidDateOrder;
